+++ xbmc/cores/DllLoader/dll_tracker_socket.cpp
@@ -5,7 +5,7 @@
 #include "dll_tracker.h"
 #include "exports/emu_socket.h"
 
-extern "C" void tracker_socket_track(unsigned long caller, SOCKET socket)
+extern "C" void tracker_socket_track(uintptr_t caller, SOCKET socket)
 {
   DllTrackInfo* pInfo = tracker_get_dlltrackinfo(caller);
   if (pInfo)
@@ -14,7 +14,7 @@ extern "C" void tracker_socket_track(unsigned long caller, SOCKET socket)
   }
 }
 
-extern "C" void tracker_socket_free(unsigned long caller, SOCKET socket)
+extern "C" void tracker_socket_free(uintptr_t caller, SOCKET socket)
 {
   DllTrackInfo* pInfo = tracker_get_dlltrackinfo(caller);
   if (pInfo)
@@ -43,7 +43,7 @@ extern "C"
 {
   SOCKET __stdcall track_socket(int af, int type, int protocol)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     SOCKET socket = dllsocket(af, type, protocol);
     if(socket>=0)
@@ -53,7 +53,7 @@ extern "C"
 
   int __stdcall track_closesocket(SOCKET socket)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     tracker_socket_free(loc, socket);
     return dllclosesocket(socket);
@@ -61,7 +61,7 @@ extern "C"
   
   SOCKET __stdcall track_accept(SOCKET s, struct sockaddr FAR * addr, OUT int FAR * addrlen)
   {
-    unsigned loc = (unsigned)_ReturnAddress();
+    uintptr_t loc = (uintptr_t)_ReturnAddress();
     
     SOCKET socket = dllaccept(s, addr, addrlen);
     if (socket>=0)
